+++ spec/spec_helper.rb
@@ -5,6 +5,13 @@ require_relative "support/components_helper"
 require "gds_api/test_helpers/content_store"
 require "faker"
 
+if ENV["USE_I18N_COVERAGE"]
+  require "i18n/coverage"
+  require "i18n/coverage/printers/file_printer"
+  I18n::Coverage.config.printer = I18n::Coverage::Printers::FilePrinter
+  I18n::Coverage.start
+end
+
 WebMock.disable_net_connect!(allow_localhost: true)
 
 RSpec.configure do |config|